smokeview source: use more digits to compute intermediate terms for colorbar labels
Former-commit-id: 7af30779099e9d913e6e7ed8d69afe98f147dcd7 [formerly 61045957ace1ae1cf5d17a69944613835720d42b]
Former-commit-id: bec8f870bd49c1f2da701155afdabea4062a8d36
Former-commit-id: 686fdb70d871140f7578b11181d901051c16291c

+++ SMV_5/source/smokeview/smoothlabel.c
@@ -10,18 +10,18 @@
 char smoothlabel_revision[]="$Revision$";
 
 void smoothlabel(float *a, float *b, int n){
-  float delta, factor, logdelta;
+  double delta, factor, logdelta;
   int ndigits;
 
-  delta = (*b-*a)/(n-2);
+  delta = ((double)*b-(double)*a)/(double)(n-2);
   if(delta==0.0)return;
   logdelta = log10((double)delta);
   ndigits=logdelta-1;
   if(logdelta<=1)ndigits--;
   factor = 5*pow(10,ndigits);
-  delta = (int)(delta/factor + 0.5f)*factor;
+  delta = (int)(delta/factor + 0.5d)*factor;
 
-  *a = factor*(int)(*a/factor+0.5f);
+  *a = factor*(int)(*a/factor+0.5d);
   *b = *a + (n-2)*delta;
 
 }
